+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Near-palindromes".
 - Tests for "Reversed look-and-say".
 - Tests for "Holey times".
+- Tests for "Consecutive balanced substrings".
 
 ## [1.15.0] - 2022-11-30
 ### Added